+++ projects/ngx-form-object-sample-app/src/app/models/user.form-object.ts
@@ -1,13 +1,8 @@
 import { Injector } from '@angular/core';
 import { Validators } from '@angular/forms';
 import { FormObject, FormObjectOptions } from 'ngx-form-object';
-import { UserService } from '../services/user/user.service';
 
 export class UserFormObject extends FormObject {
-  public serviceMappings = {
-    User: this.injector.get(UserService),
-  };
-
   public validators = {
     name: [Validators.required],
   };